Add new resource type: om_node
This resource type can be used to control nodes for the monitoring software
HP Operations Manager.  It is mostly modeled after the opcnode command.
It can be used to describe a node with assigned node groups and layout
groups.

Sample usage:

    om_node { 'test.example.com':
      label              => 'test',
      ipaddress          => '10.0.0.1',
      network_type       => 'NETWORK_IP',
      machine_type       => 'MACH_BBC_LX26RPM_X64',
      communication_type => 'COMM_BBC',
      node_type          => 'CONTROLLED',
      dynamic_ip         => no,
      layout_groups      => [
        'Linux/Webserver',
        '/CustomHierarchy/physical/rz10',
      ],
      node_groups        => [
        'linux',
        'webserver',
        'basic',
      ],
    }

The provider works with a helper script to get the current node
information. This helper script depends on perl (HP Operations Manager
shipps its own perl version and the script expects the perl executable
at the default installation directory) and it need the OVO API with perl
binding installed.