Add om_dbspi_database type
HP offers so called SmartPlugins (SPIs). One of these SPI is the DBSPI
to monitor databases. To monitor a specific database you have to create
a local configuration file with the database name, location of the
alterlog and possible custom filters.

Add a type to generate such a configration.

Usage:

    om_dbspi_database { 'TESTDB':
      username => 'monitor',
      password => 'monitor',
      logfile  => '/path/to/alert.log',
      filter   => [
        '16:content != \'UNDO\'',
        '210:content != \'UNDO\'',
      ],
    }