Replace username and password with connect property
To connect to a database the dbspi needs a connect string. This does not
necessarily mean just a username and password but can include the
hostname, port and database name as well. So instead of specifying all
the individual parts as different properties there is now only one
property called "connect"