Add dbspicfg provider for om_dbspi_database type
The dbspicfg provider uses the dbspicfg binary that is deliviered as an
instrumentation when deploying the dbspi policies to a database node.

Normally you would execute dbspicfg.sh to open up an editor and modify
the current configuration.  The dbspicfg binary on the other hand
understands the arguments -e (export) and -i (import) so we can

    1) extract the configuration and parse it
    2) modifiy it in memory
    3) generate a temporary file for the new configuration (might include
       plain text passwords!)
    4) load the new configuration
    5) remove the temp file

That is how the provider currently works