Enforce type and home property for dbspi_database type
To generate the final dbspi configuration we have to know the home
directory and the type of the database.  In case the resource describes
a database that is not yet present in the dbspi configuration we rely on
the properties the user specified in his manifest.

Solution: Make home and type property mandatory and edit the specs
accordingly