FIX errors when /var/opt/OV/bin/instrumentation is not in path
The dbspicfg executable of the dbspicfg provider may throw errors during
the first configuration when /var/opt/OV/bin/instrumentation is not in
your PATH environment.

Explicitly set PATH=/var/opt/OV/bin/instrumentation before the
dbspicfg -e call. This seems to only be an issue on the first invocation.